refactor(core): B3 follow-up — NextInt overflow, drop unused kind param, restore comments, doc+test for all-owned treasure

+++ src/Core/Random/SequentialRng.cs
@@ -22,7 +22,7 @@ public sealed class SequentialRng : IRng
     public int NextInt(int minInclusive, int maxExclusive)
     {
         if (maxExclusive <= minInclusive) return minInclusive;
-        int range = maxExclusive - minInclusive;
+        long range = (long)maxExclusive - minInclusive;
         return minInclusive + (int)(Next() % (ulong)range);
     }
